+++ .github/workflows/branch-build.yaml
@@ -199,8 +199,8 @@ jobs:
         run: |
           make release-latest-binaries
 
-  localhub-tests:
-    name: Local Hub tests
+  local-hub-tests:
+    name: Local Hub integration tests
     runs-on: ubuntu-latest
     needs: [ entry-tests, build-app, build-tests ]
     permissions:
@@ -211,11 +211,11 @@ jobs:
       - name: Setup environment
         run: |
           . ./hack/ci/setup-env.sh
-      - name: Run local Hub tests
+      - name: Run local Hub integration tests
         env:
           BUILD_IMAGES: "false"
         run: |
-          make test-localhub
+          make test-local-hub
 
   integration-tests:
     name: Integration tests
@@ -304,7 +304,7 @@ jobs:
   slackNotification:
     name: Slack Notification
     runs-on: ubuntu-latest
-    needs: [ update-cluster, localhub-tests, integration-tests ]
+    needs: [ update-cluster, local-hub-tests, integration-tests ]
     if: failure()
     steps:
       - name: Slack Notification
